# [FIX] base: improve performance on res.device for unfiltered searches