Issue:
------
Commit 749e0af3639d243f1d24e40f9c0d374c634eaf79 improves the performance of the `res.device` model in the case of a filtered search (for a given user or a given id).

However, when the search is not filtered (`SELECT * FROM res_device`), it takes a long time.
This scenario occurs when an administrator goes to the list view of this model.

Solution:
---------
For each row in the main `res_device_log` table, the subquery checks to see if a more recent (and unrevoked) `res_device_log` exists. If no similar log is found for a row in the second table then the row in the main table is considered to be the most recent.
The subquery therefore returns a boolean (`NOT EXISTS`) depending on whether or not a more recent log is present.
This boolean is assigned to a column called `is_latest_res_device_log`. Finally, we keep only the `res_device_log` in the main table which are not revoked and sort them according to `last_activity`.

This solution allows us to retain the behaviour we were looking for in the previous commit (mentioned above), i.e. to be able to use the indexes on the `user_id` and on the primary key.

In addition, the composite index is no longer necessary.

Appendices:
------------
For the query: `SELECT * FROM res_device`

Before:
```
Sort  (cost=641167.54..646433.53 rows=2106398 width=154) (actual time=24593.801..24909.190 rows=2355405 loops=1)
Sort Key: d.last_activity DESC
Sort Method: external merge  Disk: 347992kB
->  Merge Anti Join  (cost=1.11..419928.90 rows=2106398 width=154) (actual time=152.329..22591.230 rows=2355405 loops=1)
     Merge Cond: ((d.user_id = d2.user_id) AND ((d.session_identifier)::text = (d2.session_identifier)::text))
     Join Filter: ((NOT ((d2.platform)::text IS DISTINCT FROM (d.platform)::text)) AND (NOT ((d2.browser)::text IS DISTINCT FROM (d.browser)::text)) AND ((d2.last_activity > d.last_activity) OR ((d2.last_activity = d.last_activity) AND (d2.id > d.id))))
     Rows Removed by Join Filter: 64033675
     ->  Index Scan using res_device_log__composite_idx on res_device_log d  (cost=0.56..230724.16 rows=4674167 width=154) (actual time=1.367..8816.957 rows=4679105 loops=1)
     ->  Index Only Scan using res_device_log__composite_idx on res_device_log d2  (cost=0.56..142427.83 rows=4674167 width=73) (actual time=0.023..7622.788 rows=64033676 loops=1)
           Heap Fetches: 4281591
Planning Time: 1.008 ms

Execution Time: 25007.378 ms
```

After:
```
Index Scan Backward using res_device_log__last_activity_index on res_device_log d  (cost=0.43..195099.12 rows=4674167 width=154) (actual time=2.286..2091.532 rows=4679105 loops=1)
Filter: (NOT revoked)
Rows Removed by Filter: 6018
Planning Time: 0.117 ms

Execution Time: 2231.288 ms
```