@robodoo 2024-12-02
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/189281.png)](https://mergebot.odoo.com/odoo/odoo/pull/189281)